pyproject.toml
[tool.poetry]
name = "Grammar_correction_2024"
version = "0.1.0"
description = "A project for Grammar correction 2024"
authors = ["Dinh Minh Toan <toandinh6501@outlook.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.10.*"
streamlit = "^1.36.0"
google-generativeai = "^0.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"